+++ cmd/bnscli/cmd_cash_test.go
@@ -4,7 +4,6 @@ import (
 	"bytes"
 	"testing"
 
-	"github.com/iov-one/weave/cmd/bnsd/app"
 	"github.com/iov-one/weave/coin"
 	"github.com/iov-one/weave/weavetest/assert"
 	"github.com/iov-one/weave/x/cash"
@@ -22,8 +21,8 @@ func TestCmdSendTokensHappyPath(t *testing.T) {
 		t.Fatalf("cannot create a new token transfer transaction: %s", err)
 	}
 
-	var tx app.Tx
-	if err := tx.Unmarshal(output.Bytes()); err != nil {
+	tx, _, err := readTx(&output)
+	if err != nil {
 		t.Fatalf("cannot unmarshal created transaction: %s", err)
 	}