+++ cmd/bnscli/cmd_distribution.go
@@ -38,11 +38,7 @@ Create a transaction for reseting a revenue stream with a new configuration.
 			},
 		},
 	}
-	raw, err := tx.Marshal()
-	if err != nil {
-		return fmt.Errorf("cannot serialize transaction: %s", err)
-	}
-	_, err = output.Write(raw)
+	_, err = writeTx(output, tx)
 	return err
 }