+++ cmd/bnscli/cmd_distribution_test.go
@@ -5,7 +5,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/iov-one/weave/cmd/bnsd/app"
 	"github.com/iov-one/weave/weavetest/assert"
 	"github.com/iov-one/weave/x/distribution"
 )
@@ -24,9 +23,9 @@ seq:foo/bar/3,20`))
 		t.Fatalf("cannot create a transaction: %s", err)
 	}
 
-	var tx app.Tx
-	if err := tx.Unmarshal(output.Bytes()); err != nil {
-		t.Fatalf("cannot unmarshal created transaction: %s", err)
+	tx, _, err := readTx(&output)
+	if err != nil {
+		t.Fatalf("cannot read created transaction: %s", err)
 	}
 
 	txmsg, err := tx.GetMsg()