+++ cmd/bnscli/cmd_view.go
@@ -2,11 +2,9 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"flag"
 	"fmt"
 	"io"
-	"io/ioutil"
 
 	"github.com/iov-one/weave/cmd/bnsd/app"
 	"github.com/iov-one/weave/x/gov"
@@ -24,35 +22,30 @@ kind of operation are you authorizing.
 	}
 	fl.Parse(args)
 
-	raw, err := ioutil.ReadAll(input)
-	if err != nil {
-		return fmt.Errorf("cannot read transaction: %s", err)
-	}
-	if len(raw) == 0 {
-		return errors.New("no input data")
-	}
+	for {
+		tx, _, err := readTx(input)
+		if err != nil {
+			if err == io.EOF {
+				return nil
+			}
+			return err
+		}
 
-	var tx app.Tx
-	if err := tx.Unmarshal(raw); err != nil {
-		return fmt.Errorf("cannot deserialize transaction: %s", err)
+		// Protobuf compiler is exposing all attributes as JSON as
+		// well. This will produce a beautiful summary.
+		pretty, err := json.MarshalIndent(tx, "", "\t")
+		if err != nil {
+			return fmt.Errorf("cannot JSON serialize: %s", err)
+		}
+		_, _ = output.Write(pretty)
+
+		// When printing a transaction of a proposal, the embeded in proposal
+		// message is obfuscated. Extract it and print additionally.
+		_ = printProposalMsg(output, tx)
 	}
-
-	// Protobuf compiler is exposing all attributes as JSON as
-	// well. This will produce a beautiful summary.
-	pretty, err := json.MarshalIndent(tx, "", "\t")
-	if err != nil {
-		return fmt.Errorf("cannot JSON serialize: %s", err)
-	}
-	_, _ = output.Write(pretty)
-
-	// When printing a transaction of a proposal, the embeded in proposal
-	// message is obfuscated. Extract it and print additionally.
-	_ = printProposalMsg(output, tx)
-
-	return nil
 }
 
-func printProposalMsg(output io.Writer, tx app.Tx) error {
+func printProposalMsg(output io.Writer, tx *app.Tx) error {
 	msg, err := tx.GetMsg()
 	if err != nil {
 		return fmt.Errorf("cannot get transaction message: %s", err)